Add Events component and update routing in AppRoutes
